\name{simpleKIMsh-package}
\alias{simpleKIMsh-package}
\alias{simpleKIMsh}
\docType{package}
\title{
\packageTitle{simpleKIMsh}
}
\description{
\packageDescription{simpleKIMsh}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simpleKIMsh}
\packageIndices{simpleKIMsh}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{simpleKIMsh}

Maintainer: \packageMaintainer{simpleKIMsh}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
